Refactor protocol handling
Introduce UN/PACK macros as a replacement for functions of the same name. These
macros abstract away the low level details of packing and unpacking protocol
fields and make the protocol handling code easy to read and write.

As a bonus, GCC is able to generate more optimized and compact machine code
thanks to the usage of these macros.